Moved language dropdown to its own component
